move cli.init to stonks.init
allows to properly initialize the manager with the app
more semantically correct
